+++ StreamVideoTests/Utilities/Dummy/RawRecordingSettingsResponse+Dummy.swift
@@ -1,0 +1,14 @@
+//
+// Copyright © 2026 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+import StreamVideo
+
+extension RawRecordingSettingsResponse {
+    static func dummy(
+        mode: RawRecordingSettingsResponseMode = .available
+    ) -> RawRecordingSettingsResponse {
+        .init(mode: mode)
+    }
+}